Change last details and add same behaviors to recipient selector
