+++ VergeiOS/Storyboards/Transactions.storyboard
@@ -130,19 +130,11 @@
                         <viewLayoutGuide key="safeArea" id="TiJ-3y-v9e"/>
                     </view>
                     <navigationItem key="navigationItem" id="1YU-p8-wKj">
-                        <rightBarButtonItems>
-                            <barButtonItem tag="2" systemItem="trash" id="TIi-jk-9bS">
-                                <color key="tintColor" name="VergeRed"/>
-                                <connections>
-                                    <action selector="deleteContact:" destination="Bnx-KG-5UX" id="qaG-dV-5Mv"/>
-                                </connections>
-                            </barButtonItem>
-                            <barButtonItem tag="1" systemItem="save" id="erg-I3-yEF">
-                                <connections>
-                                    <action selector="saveContact:" destination="Bnx-KG-5UX" id="QwT-O7-mh0"/>
-                                </connections>
-                            </barButtonItem>
-                        </rightBarButtonItems>
+                        <barButtonItem key="rightBarButtonItem" tag="1" systemItem="save" id="erg-I3-yEF">
+                            <connections>
+                                <action selector="saveContact:" destination="Bnx-KG-5UX" id="QwT-O7-mh0"/>
+                            </connections>
+                        </barButtonItem>
                     </navigationItem>
                     <connections>
                         <segue destination="Igw-3T-6sZ" kind="show" identifier="TransactionTableViewController" id="3Na-Sc-yfM"/>
@@ -161,7 +153,7 @@
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="Dcc-rT-16c">
-                                <rect key="frame" x="0.0" y="64" width="375" height="193"/>
+                                <rect key="frame" x="0.0" y="20" width="375" height="193"/>
                                 <subviews>
                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="-" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Zge-7U-fdr">
                                         <rect key="frame" x="8" y="8" width="359" height="21"/>
@@ -191,6 +183,9 @@
                                                 </constraints>
                                                 <color key="tintColor" name="SecondaryDark"/>
                                                 <state key="normal" image="AddAddress"/>
+                                                <connections>
+                                                    <segue destination="Bnx-KG-5UX" kind="show" identifier="ContactTableViewController" id="rnG-3b-avf"/>
+                                                </connections>
                                             </button>
                                             <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Who?" textAlignment="right" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="hnr-1l-LV6">
                                                 <rect key="frame" x="28" y="0.0" width="54.5" height="28"/>
@@ -232,7 +227,7 @@
                                 </constraints>
                             </view>
                             <tableView clipsSubviews="YES" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" alwaysBounceVertical="YES" dataMode="prototypes" style="grouped" separatorStyle="default" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="18" sectionFooterHeight="18" translatesAutoresizingMaskIntoConstraints="NO" id="Ljw-hq-7uL">
-                                <rect key="frame" x="0.0" y="257" width="375" height="410"/>
+                                <rect key="frame" x="0.0" y="213" width="375" height="454"/>
                                 <color key="backgroundColor" name="BackgroundGrey"/>
                                 <prototypes>
                                     <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="none" indentationWidth="10" reuseIdentifier="transactionDetailCell" textLabel="KCn-hn-gTE" detailTextLabel="aec-BM-qFs" style="IBUITableViewCellStyleSubtitle" id="rA0-Me-VIJ">
@@ -327,12 +322,9 @@
         <namedColor name="VergeGrey">
             <color red="0.60799998044967651" green="0.60799998044967651" blue="0.60799998044967651" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
         </namedColor>
-        <namedColor name="VergeRed">
-            <color red="1" green="0.32156862745098042" blue="0.32156862745098042" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
-        </namedColor>
     </resources>
     <inferredMetricsTieBreakers>
         <segue reference="3Na-Sc-yfM"/>
-        <segue reference="hUW-WV-KOl"/>
+        <segue reference="rnG-3b-avf"/>
     </inferredMetricsTieBreakers>
 </document>
