+++ VergeiOS/Views/Send/SelectRecipientTableViewController.swift
@@ -8,56 +8,25 @@
 
 import UIKit
 
-class SelectRecipientTableViewController: EdgedTableViewController {
-
-    @IBOutlet weak var searchBar: UISearchBar!
+class SelectRecipientTableViewController: AbstractContactsTableViewController {
     
     var sendTransactionDelegate: SendTransactionDelegate!
-
-    var addresses: [Address] = []
     var sendTransaction: SendTransaction?
-    var searchQuery: String = ""
-
-    var filteredAddresses: [Address] {
-        if searchQuery == "" {
-            return addresses
-        }
-
-        return addresses.filter { (address: Address) -> Bool in
-            return address.address.lowercased().contains(self.searchQuery.lowercased())
-                || address.name.lowercased().contains(self.searchQuery.lowercased())
-        }
-    }
     
     override func viewDidLoad() {
         super.viewDidLoad()
 
-        scrollViewEdger.hideBottomShadow = true
-
         sendTransaction = sendTransactionDelegate.getSendTransaction()
-        
-        addresses = AddressBookManager().all()
-    }
 
-    // MARK: - Table view data source
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return 1
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return self.filteredAddresses.count
+        loadContacts()
     }
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "addressBookCell", for: indexPath)
+        let cell = super.tableView(tableView, cellForRowAt: indexPath)
 
-        cell.textLabel?.text = filteredAddresses[indexPath.row].name
-        cell.detailTextLabel?.text = filteredAddresses[indexPath.row].address
-        
-        if filteredAddresses[indexPath.row].address == sendTransaction?.address {
+        let address = contact(byIndexpath: indexPath)
+
+        if address.address == sendTransaction?.address {
             cell.accessoryType = .checkmark
         }
 
@@ -65,7 +34,7 @@ class SelectRecipientTableViewController: EdgedTableViewController {
     }
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        sendTransaction?.address = filteredAddresses[indexPath.row].address
+        sendTransaction?.address = contact(byIndexpath: indexPath).address
         
         sendTransactionDelegate.didChangeSendTransaction(sendTransaction!)
         
@@ -78,15 +47,5 @@ class SelectRecipientTableViewController: EdgedTableViewController {
 
     @IBAction func closeViewController(_ sender: Any) {
         self.dismiss(animated: true, completion: nil)
-    }
-}
-
-extension SelectRecipientTableViewController: UISearchBarDelegate {
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        searchQuery = searchText
-
-        DispatchQueue.main.async {
-            self.tableView.reloadData()
-        }
     }
 }
