+++ VergeiOS/Views/Transactions/ContactsTableViewController.swift
@@ -8,22 +8,7 @@
 
 import UIKit
 
-class ContactsTableViewController: EdgedTableViewController {
-
-    @IBOutlet weak var searchBar: UISearchBar!
-
-    var contacts: [[Address]] = []
-    var letters:[String] = []
-    var searchQuery: String = ""
-
-    var addressBookManager: AddressBookManager!
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        scrollViewEdger.hideBottomShadow = true
-
-        addressBookManager = AddressBookManager()
-    }
+class ContactsTableViewController: AbstractContactsTableViewController {
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
@@ -31,70 +16,6 @@ class ContactsTableViewController: EdgedTableViewController {
         loadContacts()
 
         tableView.reloadData()
-    }
-
-    func loadContacts() {
-        contacts.removeAll()
-        letters.removeAll()
-
-        let addresses = addressBookManager.all().filter { address in
-            if self.searchQuery == "" {
-                return true
-            }
-
-            return address.address.lowercased().contains(self.searchQuery.lowercased())
-                || address.name.lowercased().contains(self.searchQuery.lowercased())
-        }
-
-        let items = Dictionary(grouping: addresses, by: {
-            return String($0.name).first?.description ?? ""
-        }).sorted(by: { $0.key < $1.key })
-
-        for item in items {
-            contacts.append(item.value)
-            letters.append(item.key)
-        }
-    }
-
-    func sectionLetter(bySection section: Int) -> String {
-        return letters[section]
-    }
-
-    func contacts(bySection section: Int) -> [Address] {
-        return contacts[section]
-    }
-
-    func contact(byIndexpath indexPath: IndexPath) -> Address {
-        let items = contacts(bySection: indexPath.section)
-
-        return items[indexPath.row]
-    }
-
-    // MARK: - Table view data source
-
-    override func numberOfSections(in tableView: UITableView) -> Int {
-        // #warning Incomplete implementation, return the number of sections
-        return contacts.count
-    }
-
-    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        // #warning Incomplete implementation, return the number of rows
-        return contacts(bySection: section).count
-    }
-
-    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "addressBookCell", for: indexPath)
-
-        let address = contact(byIndexpath: indexPath)
-
-        cell.textLabel?.text = address.name
-        cell.detailTextLabel?.text = address.address
-
-        return cell
-    }
-
-    override func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        return sectionLetter(bySection: section)
     }
 
     // Override to support conditional editing of the table view.
@@ -140,15 +61,4 @@ class ContactsTableViewController: EdgedTableViewController {
         }
     }
 
-}
-
-extension ContactsTableViewController: UISearchBarDelegate {
-    func searchBar(_ searchBar: UISearchBar, textDidChange searchText: String) {
-        searchQuery = searchText
-
-        DispatchQueue.main.async {
-            self.loadContacts()
-            self.tableView.reloadData()
-        }
-    }
 }
